Move clifford synthesis tests to separate file in test/synthesis.
