+++ test/python/quantum_info/operators/symplectic/test_clifford.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2017, 2023.
+# (C) Copyright IBM 2017, 2024.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -57,12 +57,6 @@ from qiskit.quantum_info import random_clifford
 from qiskit.quantum_info.operators import Clifford, Operator
 from qiskit.quantum_info.operators.predicates import matrix_equal
 from qiskit.quantum_info.operators.symplectic.clifford_circuits import _append_operation
-from qiskit.synthesis.clifford import (
-    synth_clifford_full,
-    synth_clifford_ag,
-    synth_clifford_bm,
-    synth_clifford_greedy,
-)
 from qiskit.synthesis.linear import random_invertible_binary_matrix
 from test import QiskitTestCase  # pylint: disable=wrong-import-order
 from test import combine  # pylint: disable=wrong-import-order
@@ -586,92 +580,6 @@ class TestCliffordGates(QiskitTestCase):
         }
         expected_clifford = Clifford.from_dict(expected_clifford_dict)
         self.assertEqual(combined_clifford, expected_clifford)
-
-
-@ddt
-class TestCliffordSynthesis(QiskitTestCase):
-    """Test Clifford synthesis methods."""
-
-    @staticmethod
-    def _cliffords_1q():
-        clifford_dicts = [
-            {"stabilizer": ["+Z"], "destabilizer": ["-X"]},
-            {"stabilizer": ["-Z"], "destabilizer": ["+X"]},
-            {"stabilizer": ["-Z"], "destabilizer": ["-X"]},
-            {"stabilizer": ["+Z"], "destabilizer": ["+Y"]},
-            {"stabilizer": ["+Z"], "destabilizer": ["-Y"]},
-            {"stabilizer": ["-Z"], "destabilizer": ["+Y"]},
-            {"stabilizer": ["-Z"], "destabilizer": ["-Y"]},
-            {"stabilizer": ["+X"], "destabilizer": ["+Z"]},
-            {"stabilizer": ["+X"], "destabilizer": ["-Z"]},
-            {"stabilizer": ["-X"], "destabilizer": ["+Z"]},
-            {"stabilizer": ["-X"], "destabilizer": ["-Z"]},
-            {"stabilizer": ["+X"], "destabilizer": ["+Y"]},
-            {"stabilizer": ["+X"], "destabilizer": ["-Y"]},
-            {"stabilizer": ["-X"], "destabilizer": ["+Y"]},
-            {"stabilizer": ["-X"], "destabilizer": ["-Y"]},
-            {"stabilizer": ["+Y"], "destabilizer": ["+X"]},
-            {"stabilizer": ["+Y"], "destabilizer": ["-X"]},
-            {"stabilizer": ["-Y"], "destabilizer": ["+X"]},
-            {"stabilizer": ["-Y"], "destabilizer": ["-X"]},
-            {"stabilizer": ["+Y"], "destabilizer": ["+Z"]},
-            {"stabilizer": ["+Y"], "destabilizer": ["-Z"]},
-            {"stabilizer": ["-Y"], "destabilizer": ["+Z"]},
-            {"stabilizer": ["-Y"], "destabilizer": ["-Z"]},
-        ]
-        return [Clifford.from_dict(i) for i in clifford_dicts]
-
-    def test_decompose_1q(self):
-        """Test synthesis for all 1-qubit Cliffords"""
-        for cliff in self._cliffords_1q():
-            with self.subTest(msg=f"Test circuit {cliff}"):
-                target = cliff
-                value = Clifford(cliff.to_circuit())
-                self.assertEqual(target, value)
-
-    @combine(num_qubits=[2, 3])
-    def test_synth_bm(self, num_qubits):
-        """Test B&M synthesis for set of {num_qubits}-qubit Cliffords"""
-        rng = np.random.default_rng(1234)
-        samples = 50
-        for _ in range(samples):
-            circ = random_clifford_circuit(num_qubits, 5 * num_qubits, seed=rng)
-            target = Clifford(circ)
-            value = Clifford(synth_clifford_bm(target))
-            self.assertEqual(value, target)
-
-    @combine(num_qubits=[2, 3, 4, 5])
-    def test_synth_ag(self, num_qubits):
-        """Test A&G synthesis for set of {num_qubits}-qubit Cliffords"""
-        rng = np.random.default_rng(1234)
-        samples = 50
-        for _ in range(samples):
-            circ = random_clifford_circuit(num_qubits, 5 * num_qubits, seed=rng)
-            target = Clifford(circ)
-            value = Clifford(synth_clifford_ag(target))
-            self.assertEqual(value, target)
-
-    @combine(num_qubits=[1, 2, 3, 4, 5])
-    def test_synth_greedy(self, num_qubits):
-        """Test greedy synthesis for set of {num_qubits}-qubit Cliffords"""
-        rng = np.random.default_rng(1234)
-        samples = 50
-        for _ in range(samples):
-            circ = random_clifford_circuit(num_qubits, 5 * num_qubits, seed=rng)
-            target = Clifford(circ)
-            value = Clifford(synth_clifford_greedy(target))
-            self.assertEqual(value, target)
-
-    @combine(num_qubits=[1, 2, 3, 4, 5])
-    def test_synth_full(self, num_qubits):
-        """Test synthesis for set of {num_qubits}-qubit Cliffords"""
-        rng = np.random.default_rng(1234)
-        samples = 50
-        for _ in range(samples):
-            circ = random_clifford_circuit(num_qubits, 5 * num_qubits, seed=rng)
-            target = Clifford(circ)
-            value = Clifford(synth_clifford_full(target))
-            self.assertEqual(value, target)
 
 
 @ddt
